pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "guardrail-testing-platform"
version = "0.1.0"
description = "Agentic AI application for testing and validating LLM guardrails"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "duckduckgo-search>=7.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "click>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
]

[project.scripts]
guardrail-tester = "guardrail_tester.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/guardrail_tester"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
